Try and fix FOUC on Firefox
On Firefox browsers with a slow network, the whole website appears without CSS during a fraction of a second. That's ugly.

According to a bugzilla that I forgot to link and it's too late for that, they recommend adding a dummy script tag right after the opening of the body tag. With that, it magically fixes itself.